+++ src/main/kotlin/mb/gitonium/GitoniumPlugin.kt
@@ -74,11 +74,15 @@ class GitoniumPlugin : Plugin<Project> {
      */
     private fun registerAssertNotDirtyTask(project: Project) {
         val assertNotDirty = project.tasks.register("assertNotDirty") {
-            if (project.version.toString().endsWith(".dirty")) {
-                throw GradleException("Cannot publish a dirty version: ${project.version}")
             group = "Verification"
             description = "Asserts that there are no uncommitted changes in $project."
 
+            doLast {
+                if (project.version.toString().endsWith("+dirty")) {
+                    throw GradleException("Cannot publish a dirty version: ${project.version}")
+                }
+            }
+        }
         project.pluginManager.withPlugin("maven-publish") {
             project.tasks.named("publish") {
                 dependsOn(assertNotDirty)